New Container: Galaxy-Slurm-Node-Discovery
Unfortunately, we can't use container names as a
hostname for the Slurm nodes. Therefore, we need to
rewrite container names to the proper hostname (which
is the container id in case of Docker).
The new container handles this rewriting, by translating
the container name to its coresponding id, using the
Docker API. It will wait for the Galaxy-Configurator
to configure slurm.conf and rewrites the container names
after that.