Timeout within docker-compose tests
A simple alpine container is used for creating a timeout,
after which the setup is terminated (if `--exit-code-from`
is set).
Github Actions also offer a timeout, however using
a built-in timeout makes it possible to make a retry run,
which may be handy with flaky tests.
The default timeout is 120 minutes, but can be tweaked
using the `TIMEOUT` environment variable.