Increase Nginx timeout for tests
Some workflow tests use older tools, which are not
always available (at least it seems so). If a timeout
happens, Planemo just continues to check for the not
available tool to be installed, which results in the whole
test to time out. This change increases the UWSGI read
timeout of Nginx to 60 minutes to help with that.

* Introducing NGINX_UWSGI_READ_TIMEOUT
* Setting NGINX_UWSGI_READ_TIMEOUT to 3600s for
tests