Include parts of galaxy/lib into HTCondor and Slurm image
To set the metadata, the worker nodes need the  `galaxy_ext` library and parts of `galaxy`, which earlier
was not available to
them. Instead of exporting it from Galaxy directly (which
may result in degraded performance for the whole setup
on Mac because
of the shared filesystem), this change includes the library
directly in the image.

This should fix the following error:

> from galaxy_ext.metadata.set_metadata import set_metadata; set_metadata() ImportError: No module named galaxy_ext.metadata.set_metadata