Use HTCondor config from configurator
The HTCondor containers mount
`$EXPORT_DIR/htcondor` and copy the correct
config file (determined by the env `HTCONDOR_TYPE`)
to /etc/condor/condor_config.local for HTCondor to use.
Before that, it waits for the configurator to release the
lock, to avoid race condition.

For Galaxy, the HTCondor config is mounted under
`/htcondor_config`. The start script will check this folder
to determine whethter it should start HTCondor and
copies the config to /etc/condor/condor_config.local.