Base image for cluster clients (Galaxy & Pulsar)
Installing the Slurm and HTCondor clients
is common for Galaxy and Pulsar.
To reduce code  duplication, its probably
better to put this into
its own base-image (the same approach as the
container-base).

This image contains HTConodor, Slurm, and
CVMFS, together with the inheritet contents of
the  `container-base` image (Docker &
Singularity).