Galaxy: Create default admin user AFTER startup
Earlier, we could create a user before
Galaxy started. This is not possible
anymore, as Galaxy does the DB
initialization (so before that, the DB
doesn't know anything about users etc).
With this change, we start a background
job, that will wait 'til Galaxy is ready
and then create the default user.
As we also support `GALAXY_PROXY_PREFIX`,
the configurator now creates a `GALAXY_PROXY_PREFIX.txt`
file within the GALAXY_CONFIG_DIR, that
allows the startup script to get the
correct proxy-prefix.