CI: Tests for Pull Requests
The current CI was only running when pushing
to a branch, so we couldn't easily tell if a PR
would pass all tests. As the main CI workflow
pushes to a registry (to make local debugging
and continuous development on a branch
easier), we also couldn't just reuse the whole
concept for PRs.
This new `pull-request` worfklow fixes this for
now (though not perfectly) by building and
testing in the same job (so no need to share
data between runners). This is not really
efficient, as we build all the containers for
each and every test, however it is the simplest
and most straightforward solution for now.