:recycle: Configurator: Default config dir vars
Instead of setting a default of the
`*_CONFIG_DIR` environment variables for every
usage, we should instead define the defaults
once and simplify the rest of the usages.
But as the `_CONFIG_` prefix is already
in use to set arbritary configs for Galaxy etc,
chaging it to `*_CONF_DIR` at the same time
to avoid collisions.